refactor(config): split provider model transforms
